move lz4 compression into background worker thread
