--- frontend/javascripts/oxalis/workers/byte_array_to_base64.worker.js
@@ -1,10 +1,0 @@
-// @flow
-import Base64 from "base64-js";
-
-import { expose } from "./comlink_wrapper";
-
-function byteArrayToBase64(byteArray: Uint8Array): string {
-  return Base64.fromByteArray(byteArray);
-}
-
-export default expose<typeof byteArrayToBase64>(byteArrayToBase64);
